backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
PyJWT==2.8.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
httpx==0.25.2

# AI/ML Dependencies - IMPORTANT: Use latest versions
transformers>=4.38.0
torch>=2.1.0
accelerate>=0.25.0
sentencepiece>=0.1.99
protobuf>=3.20.0